pyproject.toml
[project]
name = "masters-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.60.2",
    "openai-whisper>=20240930",
    "requests>=2.32.3",
]
